+++ addOns/client/CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Added
+- Note about custom containers in the help.
+
 ### Fixed
 - Cases where the Firefox profile was not added successfully via Selenium.