+++ addOns/client/CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+
 ## [0.4.0] - 2023-10-31
 ### Added
 - Note about custom containers in the help.