+++ addOns/client/CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this add-on will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
+## [0.5.0] - 2023-11-07
 ### Added
 - Client History and Details context menu items.
 
@@ -55,6 +55,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - First version.
 
+[0.5.0]: https://github.com/zaproxy/zap-extensions/releases/client-v0.5.0
 [0.4.0]: https://github.com/zaproxy/zap-extensions/releases/client-v0.4.0
 [0.3.0]: https://github.com/zaproxy/zap-extensions/releases/client-v0.3.0
 [0.2.0]: https://github.com/zaproxy/zap-extensions/releases/client-v0.2.0